+++ src/form-elements/FormElementFreshdeskDependentField.tsx
@@ -1,7 +1,9 @@
 import * as React from 'react'
 import { FormTypes } from '@oneblink/types'
 import FormElementForm, { Props } from './FormElementForm'
-import generateFreshdeskDependentFieldElements from '../services/generateFreshdeskDependentFieldElements'
+import generateFreshdeskDependentFieldElements, {
+  getNestedOptions,
+} from '../services/generateFreshdeskDependentFieldElements'
 
 function FormElementFreshdeskDependentField({
   element,
@@ -9,14 +11,43 @@ function FormElementFreshdeskDependentField({
 }: Omit<Props, 'element'> & {
   element: FormTypes.FreshdeskDependentFieldElement
 }) {
+  const freshdeskElements = React.useMemo(
+    () => generateFreshdeskDependentFieldElements(element),
+    [element],
+  )
+
   const formElement = React.useMemo<FormTypes.FormFormElement>(() => {
+    const [categoryElement, subcategoryElement, itemElement] = freshdeskElements
+    const elements = [categoryElement]
+    const value = props.value as
+      | FormTypes.FreshdeskDependentFieldElementValue
+      | undefined
+
+    if (props.value?.category) {
+      const subCategoryOptions = getNestedOptions(
+        element.options,
+        value?.category,
+      )
+      subcategoryElement.options = subCategoryOptions
+      elements.push(subcategoryElement)
+
+      if (value?.subCategory) {
+        const itemOptions = getNestedOptions(
+          subCategoryOptions,
+          value?.subCategory,
+        )
+        itemElement.options = itemOptions
+        elements.push(itemElement)
+      }
+    }
+
     return {
       ...element,
       type: 'form',
       formId: NaN,
-      elements: generateFreshdeskDependentFieldElements(element, props.value),
+      elements,
     }
-  }, [element, props.value])
+  }, [element, props.value, freshdeskElements])
 
   return <FormElementForm element={formElement} {...props} />
 }