+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `isDirty` validation for Checkbox Element
+- freshdesk dependent elements not validating correctly
 
 ## [2.13.0] - 2023-04-27